autenticacao/requirements.txt
# Dependências mínimas para o serviço de autenticação (copiado/adaptado do backend)
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
fastapi-mail

sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
argon2-cffi==23.1.0
python-dotenv==1.0.1

pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2
